Make code compatible with python 2.7
The instructions for using the GPIO on the onion are here:
https://docs.onion.io/omega2-docs/gpio-python-module.html

These instructions use python-light and pyOnionGpio, both of which are for
python 2.7.

There is a package called python3-light, but I don't think there is the
equivalent for the pyOnionGpio package.

Worth noting, that is sounds like while this pyOnionGpio package is the one
mentioned in the official docs, it's a "quick and dirty wrapper for the sysfs
gpio interface" according to this post:
https://community.onion.io/topic/3591/python3-modules-for-gpio-and-serial-port?_=1690687687356&lang=en-GB

Apparently any of these libraries would work:
  - https://pypi.org/project/gpio/
  - https://python-periphery.readthedocs.io/en/latest/gpio.html
  - https://github.com/vitiral/gpio

For now though, it's easy enough to use python 2.7 but the option to upgrade and
change GPIO libraries is there is we hit a snag

+++ lib/stepper.py
@@ -29,7 +29,7 @@ class Stepper:
             self._step(-1)
 
     def set_percentage(self, percentage):
-        desired_step = round(percentage * self.range)
+        desired_step = int(round(percentage * self.range))
         steps = desired_step - self.current_step
         if steps > 0:
             self.forward(steps)
